eks-pods/auth-pod/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
psycopg[binary,pool]==3.2.3
authlib==1.3.2
httpx==0.28.1
PyJWT[crypto]==2.10.1
itsdangerous==2.2.0
